move models to models/entities, modify buddyrepository
