+++ CodeBuddies/CodeBuddies.csproj
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Resources\Data\" />
+    <Folder Include="Resources\Data\Constants\" />
     <Folder Include="Services\" />
     <Folder Include="ViewModels\" />
   </ItemGroup>
