+++ CodeBuddies/Models/Entities/TextEditor.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CodeBuddies.Models
+namespace CodeBuddies.Models.Entities
 {
     internal class TextEditor
     {
@@ -50,7 +50,7 @@ namespace CodeBuddies.Models
             // Open the file
         }
 
-     
+
 
     }
 }
